requirements.txt
watchdog==2.1.6
slackclient==2.9.3
elasticsearch==8.0.0
eml-parser==1.17.0
